refactor: move styles to external file
